Commit para ajustar detalles en las pantallas anteriormente hechas
