+++ src/app/pages/administrative-processes/conversion-act/conversion-act/conversion-act.component.html
@@ -134,11 +134,11 @@
       <div class="row mt-5">
         <div class="col-md-12 col-sm-12 col-xs-12">
           <div class="justify-content-center">
-            <button class="btn btn-primary btn-sm active" (click)="detail()">
+            <button class="btn btn-info btn-sm active" (click)="detail()">
               Detalles Acta
             </button>
             &nbsp;
-            <button class="btn btn-primary btn-sm active" (click)="close()">
+            <button class="btn btn-danger btn-sm active" (click)="close()">
               Cierra Acta
             </button>
             &nbsp;
@@ -220,7 +220,7 @@
         <div body>
             <div class="col-md-12 col-sm-12 col-xs-12">
                 <div class="justify-content-center">
-                  <button class="btn btn-primary btn-sm active" (click)="add()">
+                  <button class="btn btn-success btn-sm active" (click)="add()">
                     Agregar
                   </button>
                 </div>
