+++ src/app/pages/administrative-processes/conversion-management/conversion-management/conversion-management.component.html
@@ -6,7 +6,7 @@
   <div body>
     <form [formGroup]="form" class="form-material">
       <!--Date Form-->
-      <div class="row">
+      <div class="row mt-5">
         <div class="col-md-4 col-sm-12 col-xs-12">
           <form-field
             [control]="form.get('idConversion')"
@@ -34,7 +34,7 @@
           </form-field>
         </div>
       </div>
-      <div class="row">
+      <div class="row mt-5">
         <div class="col-md-6 col-sm-12 col-xs-12">
           <form-field [control]="form.get('tipo')" label="Tipo">
             <select class="form-control" formControlName="tipo">
@@ -55,7 +55,7 @@
           </form-field>
         </div>
       </div>
-      <div class="row">
+      <div class="row mt-5">
         <div class="col-md-6 col-sm-12 col-xs-12">
           <form-field
             [control]="form.get('actaConversion')"
@@ -92,7 +92,7 @@
                   </form-field>
                 </div>
               </div>
-              <div class="row">
+              <div class="row mt-5">
                 <div class="col-md-12 col-sm-12 col-xs-12">
                     <form-field
                     [control]="form.get('description')"
@@ -106,16 +106,18 @@
               </div>
         </div>
       </app-card>
-      <div class="row">
+      <div class="row mt-5">
         <div class="col-md-6 col-sm-12 col-xs-12 mt-2">
           <div class="add-select">
             <button
               class="btn btn-primary btn-sm active"
-              >
+              (click)="generatePaswword()"
+              [disabled]="saved">
               Generar PW
             </button>
           </div>
         </div>
+        
         <div class="col-md-6 col-sm-12 col-xs-12 mt-2">
           <div class="add-select">
             <button 
@@ -126,6 +128,12 @@
           </div>
         </div>
       </div>
+
+        <div class="row mt-5" *ngIf="generarPass">
+          <div class="col-md-4 col-sm-12 col-xs-12 mt-2">
+            <p>Nueva contraseña:  <strong>{{password}}</strong></p>
+          </div>
+        </div>
     </form>
   </div>
 </app-card>
