--- src/app/pages/administrative-processes/conversion-management/password-modal/password-modal.component.html
@@ -1,30 +1,0 @@
-<app-modal>
-    <div header>
-      <h5 class="modal-title">{{ title }}</h5>
-    </div>
-    <div body>
-      <p>
-        Lorem, ipsum dolor sit amet consectetur adipisicing elit. Nihil
-        perspiciatis in impedit sunt officia aspernatur quam consequatur qui,
-        aperiam, natus ratione! Maiores dolorum magnam et accusantium
-        assumenda. Autem, enim optio.
-      </p>
-    </div>
-    <div footer>
-      <div class="d-flex justify-content-center">
-        <div class="m-3">
-          <button type="button" class="btn btn-primary active">
-            Guardar
-          </button>
-        </div>
-        <div class="m-3">
-          <button
-            type="button"
-            class="btn btn-danger active"
-            (click)="close()">
-            Cancelar
-          </button>
-        </div>
-      </div>
-    </div>
-  </app-modal>
